+++ ray-on-gke/kuberay-tpu-webhook/deployments/webhook-svc.yaml
@@ -1,0 +1,16 @@
+apiVersion: v1
+kind: Service
+metadata:
+  labels:
+    app: kuberay-tpu-webhook
+  name: kuberay-tpu-webhook
+  namespace: default
+spec:
+  type: ClusterIP
+  selector:
+    app: kuberay-tpu-webhook
+  ports:
+    - port: 443
+      targetPort: https
+      protocol: TCP
+      name: https
