--- ray-on-gke/user/kuberay-tpu-webhook/Dockerfile
@@ -1,17 +1,0 @@
-FROM google-go.pkg.dev/golang:1.21.3 AS builder
-
-WORKDIR /app
-COPY go.mod go.sum ./
-RUN go mod download -x
-
-COPY main.go ./
-
-# Purposefully set AFTER downloading and caching dependencies
-ARG TARGETOS TARGETARCH
-RUN CGO_ENABLED=0 GOOS=$TARGETOS GOARCH=$TARGETARCH go build -ldflags="-w -s" -o /go/bin/kuberay-tpu-webhook .
-
-FROM scratch
-COPY --from=builder /go/bin/kuberay-tpu-webhook /usr/local/bin/kuberay-tpu-webhook
-
-EXPOSE 443
-ENTRYPOINT ["kuberay-tpu-webhook"]
