--- ray-on-gke/user/kuberay-tpu-webhook/deployments/deployment.yaml
@@ -1,43 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  labels:
-    app: kuberay-tpu-webhook
-  name: kuberay-tpu-webhook
-  namespace: default
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: kuberay-tpu-webhook
-  template:
-    metadata:
-      labels:
-        app: kuberay-tpu-webhook
-    spec:
-      containers:
-        - image: us-docker.pkg.dev/ai-on-gke/kuberay-tpu-webhook/kuberay-tpu-webhook:latest
-          imagePullPolicy: Always
-          name: kuberay-tpu-webhook
-          ports:
-          - name: https
-            containerPort: 443
-            protocol: TCP
-          livenessProbe:
-            httpGet:
-              path: /
-              port: https
-              scheme: HTTPS
-          readinessProbe:
-            httpGet:
-              path: /
-              port: https
-              scheme: HTTPS
-          volumeMounts:
-            - name: tls
-              mountPath: "/etc/kuberay-tpu-webhook/tls"
-              readOnly: true
-      volumes:
-        - name: tls
-          secret:
-            secretName: kuberay-tpu-webhook-certs
