--- ray-on-gke/user/kuberay-tpu-webhook/tests/example-pod.yaml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: Pod
-metadata:
-  name: example-pod
-  labels:
-    app: kuberay-tpu-webhook
-    kuberay-tpu-webhook/inject: enabled
-spec:
-  containers:
-  - name: example-container
-    image: busybox
-    env: []
