--- ray-on-gke/user/kuberay-tpu-webhook/tests/ray-cluster.yaml
@@ -1,83 +1,0 @@
-apiVersion: ray.io/v1
-kind: RayCluster
-metadata:
-  name: example-cluster
-  labels: 
-    kuberay-tpu-webhook/inject: enabled
-spec:
-  rayVersion: "2.6.1-py310"
-  enableInTreeAutoscaling: true
-  headGroupSpec:
-    serviceType: ClusterIP
-    rayStartParams:
-      dashboard-host: "0.0.0.0"
-      block: 'true'
-    template:
-        metadata:
-          labels:
-            kuberay-tpu-webhook/inject: disabled
-        spec: # Pod spec
-            containers:
-            - name: ray-head
-              image: rayproject/ray:2.6.1-py310
-              resources:
-                limits:
-                  cpu: "8"
-                  memory: "40G"
-                  ephemeral-storage: 20Gi
-                requests:
-                  cpu: "8"
-                  memory: "40G"
-                  ephemeral-storage: 10Gi
-              lifecycle:
-                preStop:
-                  exec:
-                    command: ["/bin/sh","-c","ray stop"]
-              ports:
-              - containerPort: 6379
-                name: gcs
-              - containerPort: 8265
-                name: dashboard
-              - containerPort: 10001
-                name: client
-              - containerPort: 8000
-                name: serve
-  workerGroupSpecs:
-  - groupName: small-group
-    replicas: 3
-    minReplicas: 1
-    maxReplicas: 3
-    rayStartParams:
-      block: 'true'
-      resources: '"{\"google.com/tpu\": 4}"'
-    template:
-      metadata:
-        labels:
-          cloud.google.com/gke-ray-node-type: worker
-          kuberay-tpu-webhook/inject: enabled
-        annotations:
-          key: value
-      spec:
-        nodeSelector:
-          iam.gke.io/gke-metadata-server-enabled: "true"
-          cloud.google.com/gke-tpu-accelerator: tpu-v4-podslice
-          cloud.google.com/gke-tpu-topology: 2x2x1
-        containers:
-        - name: ray-worker
-          image: 'busybox:1.28'
-          env: []
-          resources:
-            limits:
-              cpu: "1"
-              google.com/tpu: "4"
-              memory: "40G"
-              ephemeral-storage: 20Gi
-            requests:
-              cpu: "1"
-              google.com/tpu: "4"
-              memory: "40G"
-              ephemeral-storage: 10Gi  
-        volumes:
-          - name: log-volume
-            emptyDir: {}
-
